changes bounded with moving dateTimer to own package
