+++ src/main/java/com/jarq/system/managers/filesManagers/RepositoryManager.java
@@ -91,7 +91,7 @@ public class RepositoryManager implements IRepositoryManager {
 //
 //            }
 //        }
-//        System.out.println("removing file or directory : " + dir.getName());
+//        System.out.println("removing filepath or directory : " + dir.getName());
 //        return dir.delete();
 
         return false;
