+++ src/main/java/com/jarq/system/models/content/SQLiteDaoContent.java
@@ -3,7 +3,7 @@ package com.jarq.system.models.content;
 import com.jarq.system.dao.SqlDao;
 import com.jarq.system.enums.DbTable;
 import com.jarq.system.exceptions.DaoFailure;
-import com.jarq.system.helpers.IDateTimer;
+import com.jarq.system.helpers.datetimer.IDateTimer;
 import com.jarq.system.helpers.repositoryPath.IRepositoryPath;
 import com.jarq.system.managers.databaseManagers.JDBCProcessManager;
